update task 0

+++ 0x07-pointers_arrays_strings/0-memset.c
@@ -9,11 +9,11 @@
  * return: a pointer to the filled memory area @s
  */
 
-void *_memset(char *s, char b, size_t n)
+void *_memset(char *s, char b, unsigned int n)
 {
 	unsigned int a;
 
-	for (a = 0; a< n; a++)
+	for (a = 0; a < n; a++)
 	{
 		s[a] = b;
 	}
